requirements.txt
python-dotenv
ollama
chromadb
llama-index-core
llama-index-embeddings-ollama
llama-index-llms-ollama
llama-index-vector-stores-chroma